refactor(portfolio): extract messages dump viewer
Move the messages dump modal into a dedicated chat component.
Keep ChatResults focused on metadata display and dialog state.
